Add ClientQueryEntityMetadataRegistrationTestJavaOnlyServer
